LayoutTests/imported/w3c/web-platform-tests/intersection-observer/client-rect-expected.txt


PASS IntersectionObserverEntry.boundingClientRect should match target.boundingClientRect() 
FAIL First rAF should generate notification. assert_equals: One notification. expected 1 but got 0
